Optimisation complete without testing.
I have already done all of the optimisation steps based on Herr
Bauernöppel source code on Gauss Filter and his notes.

I have also removed the if modulo function and added y+=2 and x+=2 to
read only the needed lines in the array.

One more thing for the last function wavelet_filter_restrict_manuell, I
don’t really know if it is correct or not, and I think this is the only
function that is most optimised, when correct.

There are notes commented in the source code. Maybe you can used it as
a guideline to do the documentation.